Package Downloader (wip/3)

+++ build-assets.sh
@@ -8,16 +8,21 @@ build_widget() {
 
     echo "$widget_dir"
     pdc "$widget_dir" "Assets/Widgets/${widget_name}.pdx"
-    echo "/Shared/FunnyOS/Widgets/${widget_name}.pdx/" > "Assets/Widgets/${widget_name}.pdx/installpath"
+    if [ ! -f "Assets/Widgets/${widget_name}.pdx/installpath" ]; then
+        echo "/Shared/FunnyOS/Widgets/${widget_name}.pdx/" > "Assets/Widgets/${widget_name}.pdx/installpath"
+    fi
 
     if [ -f "Assets/Widgets/${widget_name}.pdx.zip" ]; then
         rm "Assets/Widgets/${widget_name}.pdx.zip"
     fi
-    zip -r "Assets/Widgets/${widget_name}.pdx.zip" "Assets/Widgets/${widget_name}.pdx/"
-    rm -r "Assets/Widgets/${widget_name}.pdx/"
+    cd "Assets/Widgets"
+    zip -r "${widget_name}.pdx.zip" "${widget_name}.pdx/"
+    rm -r "${widget_name}.pdx/"
+    cd -
 }
 
 build_widget "Widgets/FunnyLoader" "FunnyLoader"
 build_widget "Widgets/To-Do List" "Todo"
 build_widget "Widgets/Badge Downloader" "BadgeDownloader"
 build_widget "Widgets/Package Downloader" "PackageDownloader"
+build_widget "Widgets/File Explorer" "FileExplorer"
